Fixed compilation and tests for Metal

+++ Tests/IncludeTest/GraphicsEngineMetal/RasterizationRateMapMtlH_test.m
@@ -43,6 +43,6 @@ void TestRasterizationRateMapMtl_CInterface(IRasterizationRateMapMtl* pRRM)
     IRasterizationRateMapMtl_GetPhysicalGranularity(pRRM, (Uint32*)NULL, (Uint32*)NULL);
     IRasterizationRateMapMtl_MapScreenToPhysicalCoordinates(pRRM, 0, 0.f, 0.f, (float*)NULL, (float*)NULL);
     IRasterizationRateMapMtl_MapPhysicalToScreenCoordinates(pRRM, 0, 0.f, 0.f, (float*)NULL, (float*)NULL);
-    IRasterizationRateMapMtl_GetParameterBufferSizeAndAlign(pRRM, (Uint32*)NULL, (Uint32*)NULL);
+    IRasterizationRateMapMtl_GetParameterBufferSizeAndAlign(pRRM, (Uint64*)NULL, (Uint32*)NULL);
     IRasterizationRateMapMtl_CopyParameterDataToBuffer(pRRM, (struct IBuffer*)NULL, 0);
 }
